Calculator and Strategy tests
