+++ pruby/gui/__init__.py
@@ -1,2 +1,3 @@
 from .gridable import FilenameEntry, UfloatEntry, StatusBar
 from .popups import open_file_dialogue, show_error, show_about
+from .app import Application
