+++ resources/example2.txt
